requirements.txt
PyMuPDF
opencv-python-headless
numpy
Pillow
pytesseract